PoC for a combined force
Here is my PoC for a combined force rather than separate ones. 
It might have some bugs in it as I whipped this all together in less than 1.5hrs of pissing around. But it should allow people to choose there detachments. The number of selectable detachments is based on how many command you have in the force, and presumes that you have a crusade primary detachment. 
It allows any of the Primes to select to have the Logistics prime to grant another unit.
It also makes sure as the command one and troops one is limited to max 1 between the 2 if you don't also have a command cadre as otherwise you would have 2 picks of logistics in a single detachment.